Refactoring the db
